% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fbm.R
\name{add_to_fbm}
\alias{add_to_fbm}
\title{Append a new chromosome to the FBM}
\usage{
add_to_fbm(
  ancestry_file,
  ancestry_fmt,
  plink_prefix,
  fbm_prefix,
  variants = NULL,
  idx_variants = NULL,
  min_ac = 0,
  samples = NULL,
  idx_samples = NULL,
  anc_names = NULL,
  chunk_size = 400,
  fbm = NULL
)
}
\arguments{
\item{ancestry_file}{A string with the file path for a single local
ancestry input file}

\item{ancestry_fmt}{A string with the local ancestry format
of \code{ancestry_file}. Either "FLARE", "RFMix", or "lanc"}

\item{plink_prefix}{A string with the prefix for a single set of plink2
files}

\item{fbm_prefix}{A string with the prefix for the
file path where a new backing file for the FBM
will be written. This prefix will be appended with ".bk"}

\item{variants}{A character vector of variant IDs to
include in FBM}

\item{min_ac}{An integer with a minimum cut off for
allele counts in each column of the FBM}

\item{samples}{A character vector of samples to
include in FBM}

\item{idx_samples}{An integer vector of sample indices,
corresponding to plink2 psam input, to include in FBM}

\item{anc_names}{A character vector containing (ordered)
names for each population in \code{ancestry_file}. If not
provided, integer values are used}

\item{chunk_size}{An integer with the max number of
variants to load at a given time while reading the
plink2 pgen file}

\item{fbm}{An existing \code{FBM.code256} object to which
results are appended.}
}
\value{
A list containing:
\code{fbm}: an \code{FBM.code256} object from \code{bigstatsr},
which links to a file-backed matrix used for
fittting HAUDI models. Each column contains
either ancestry-specific genotypes (e.g. number of
alternate alleles where ancestry is also "pop0") or
the total genotype. This object also contains a
\code{samples} attribute ordered according to the samples
in the matrix rows.
\code{info} a \code{data.table} containing information
about columns in the FBM.
}
}
\description{
Append a new chromosome to the FBM
}
