% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaudi.R
\name{construct_gaudi}
\alias{construct_gaudi}
\title{construct_gaudi}
\usage{
construct_gaudi(fbm_obj, fbm_info, snps = NULL)
}
\arguments{
\item{fbm_obj}{object of FBM class}

\item{fbm_info}{data frame containing information}

\item{snps}{vector of SNPs}
}
\value{
matrix for fitting GAUDI
}
\description{
Constructs X matrix for GAUDI
}
\author{
Frank Ockerman
}
