% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracts.R
\name{convert_to_lanc}
\alias{convert_to_lanc}
\title{Convert an "RFMix" or "FLARE" file to ".lanc" format}
\usage{
convert_to_lanc(file, file_fmt, plink_prefix, output)
}
\arguments{
\item{file}{A string with the file path for the ancestry file to be converted}

\item{file_fmt}{A string with the file format to be converted
(either "FLARE" or "RFMix")}
}
\description{
Convert an "RFMix" or "FLARE" file to ".lanc" format
}
