% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaudi.R
\name{gaudi}
\alias{gaudi}
\title{gaudi}
\usage{
gaudi(
  fbm_obj,
  fbm_info,
  y,
  gamma_vec,
  k = 10,
  ind_train = NULL,
  snps = NULL,
  verbose = FALSE,
  minlam = 0,
  maxsteps_init = 2000,
  maxsteps_cv = 2000 * 10,
  approx = FALSE
)
}
\arguments{
\item{fbm_obj}{object of FBM class}

\item{fbm_info}{data frame containing information
for fbm (chrom/pos/samples/etc.)}

\item{y}{vector of responses}

\item{gamma_vec}{vector of gamma values}

\item{k}{number of folds to use in cross-validation}

\item{ind_train}{integer vector of samples to use in training model}

\item{snps}{vector of snp names to use in model}

\item{verbose}{boolean indicating verbosity}

\item{minlam}{minimum lambda value for genlasso}

\item{maxsteps_init}{max steps for genlasso in initial model}

\item{maxsteps_cv}{max steps for genlasso in cross-validation models}

\item{approx}{boolean indicating whether}
}
\value{
list containing information for best cross-validated fit
}
\description{
gaudi
}
\author{
Bryce Rowland, Frank Ockerman
}
