% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haudi.R
\name{get_beta_haudi}
\alias{get_beta_haudi}
\title{Gets data frame with ancestry-specific effects estimated by HAUDI}
\usage{
get_beta_haudi(fbm_info, haudi_model)
}
\arguments{
\item{fbm_info}{A data.table with information on the columns in the FBM}

\item{haudi_model}{A haudi model, as returned by \code{haudi()}}
}
\description{
Gets data frame with ancestry-specific effects estimated by HAUDI
}
