% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fbm.R
\name{initialize_fbm}
\alias{initialize_fbm}
\title{Creates a new File-Backed matrix for HAUDI}
\usage{
initialize_fbm(fbm_prefix, samples)
}
\arguments{
\item{fbm_prefix}{A string with the prefix for
the intended FBM file path}
}
\value{
An FBM object with an attribute "samples"
containing an ordered vector of samples
}
\description{
This utility function creates an FBM.256 object
with the correct coding to represent genotype
values from 0 to 2.
}
