% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HAUDI.R
\name{LASSO}
\alias{LASSO}
\title{LASSO}
\usage{
LASSO(FBM_obj, FBM_info, y, gamma, ind_train = NULL, family, snps = NULL, ...)
}
\arguments{
\item{FBM_obj}{object of FBM class}

\item{FBM_info}{data frame containing information for FBM (chrom/pos/samples/etc.)}

\item{y}{Vector of responses}

\item{ind_train}{Vector of indices specifying the rows to use for training the model}

\item{family}{Either "gaussian" or "binomial"}

\item{snps}{Vector of SNPs to include in model}

\item{...}{additional arguments to pass to big_spLinReg / big_spLogReg}
}
\value{
An object of class \code{big_sp_list} from the \code{bigstatsr} package
}
\description{
LASSO
}
\author{
Frank Ockerman
}
