% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fbm.R
\name{make_haudi_chunk}
\alias{make_haudi_chunk}
\title{Produces HAUDI matrix and variant info for a single chunk}
\usage{
make_haudi_chunk(chunk, gr_tracts, pgen, pvar, min_ac, anc_names, idx_samples)
}
\arguments{
\item{chunk}{An integer vector of indices in the pgen file}

\item{gr_tracts}{A \code{GRanges} object corresponding to the ancestry file}

\item{pgen}{A \code{pgen} object produced by \code{pgenlibr::NewPgen()}}

\item{pvar}{A data frame corresponding to the pvar file}

\item{min_ac}{An integer with the minimum allele count to
retain columns in the HAUDI matrix}

\item{anc_names}{An ordered vector of ancestry names}

\item{idx_samples}{An ordered integer vector of samples to
retain from the pgen}
}
\value{
A list with the HAUDI matrix and variant info for the chunk
}
\description{
Produces HAUDI matrix and variant info for a single chunk
}
