% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rf_FBM.R
\name{make_rf_FBM}
\alias{make_rf_FBM}
\title{Make FBM containing ancestry-genotypes and return list with containing FBM and SNP info}
\usage{
make_rf_FBM(vcf_file, rf_file, FBM_pref, chunk_size, rds = NULL, minAC = 1)
}
\arguments{
\item{vcf_file}{a file path to a VCF file}

\item{rf_file}{a file path to an RFMix output file}

\item{FBM_pref}{a file path to store FBM. Omit the file extension}

\item{chunk_size}{an integer indicating the max number of VCF records to read at a time}

\item{minAC}{an integer indicating the minimum allele count (per-ancestry) to retain}
}
\description{
Make FBM containing ancestry-genotypes and return list with containing FBM and SNP info
}
\author{
Frank Ockerman
}
