% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracts.R
\name{query_tracts}
\alias{query_tracts}
\title{Query ancestry tracts for haplotype ancestry matrices}
\usage{
query_tracts(query, tracts)
}
\arguments{
\item{query}{An integer vector of indices corresponding to a
.lanc and .pvar file}

\item{tracts}{A list with ancestry tracts (as returned by \code{read_lanc})}
}
\value{
A list with two elements "hap0" and "hap1", containing the
haplotype-level ancestry matrices for the query.
}
\description{
Query ancestry tracts for haplotype ancestry matrices
}
