% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracts.R
\name{read_ancestry_tracts}
\alias{read_ancestry_tracts}
\title{Read ancestry tracts to a data.table}
\usage{
read_ancestry_tracts(
  file,
  file_fmt,
  extend_tracts = FALSE,
  plink_prefix = NULL
)
}
\arguments{
\item{file}{Path for the local ancestry input file}

\item{file_fmt}{Either "FLARE", "RFMix", or "lanc", specifying
the format of the local ancestry file.}

\item{extend_tracts}{Boolean indicating whether to
impute ancestry at before/after the minimum/maximum
position on a chromosome. This ensures that any
position on the chromosome may be queried for ancestry.
that all}
}
\value{
A data.table with one row per-ancestry tract, containing
columns "sample", "hap" (haplotype), "chrom" (chromosome), "spos"
(start position), "epos" (end position) and "ancestry" (an integer).
}
\description{
\code{read_ancestry_tracts} processes a local ancestry input file
and splits it into ancestry tracts per-sample and per-haplotype,
returning a data.table with one row for each tract.
}
\details{
This function currently supports RFMix, FLARE, or the .lanc input
format defined by \code{Admix-kit}. Multi-chromosome input files are
allowed for RFMix and FLARE but not .lanc. It is assumed that
input files are sorted by chromosome and position. For RFMix,
it is assumed that CRF points are contiguous. When the ancestry
of a haplotype switches (or when a new chromosome begins),
the haplotype is split into ancestry tracts.
For FLARE input, which is not contiguous,
ancestry is imputed to the midpoint between variants.
For .lanc input, tracts are contiguous by definition.
}
\author{
Franklin Ockerman
}
