% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracts.R
\name{read_lanc}
\alias{read_lanc}
\title{Read a ".lanc" format ancestry file}
\usage{
read_lanc(lanc_file, plink_prefix)
}
\arguments{
\item{lanc_file}{A string with the ".lanc" file path}

\item{plink_prefix}{A string with the prefix of the plink2 files}
}
\value{
A data.table
}
\description{
This utility function read ".lanc" files into a data.table
where each row is a single per-sample/haplotype ancestry tract.
}
