% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fbm.R
\name{resolve_indices}
\alias{resolve_indices}
\title{Match indices from (optional) subset and reference}
\usage{
resolve_indices(ids = NULL, idx = NULL, reference, label = "items")
}
\arguments{
\item{ids}{A character vector which contains
a subset of \code{reference}}

\item{idx}{An integer vector of indices
representing a subset of \code{reference}}

\item{reference}{A vector of values
which we want indices from}

\item{label}{A string with what the \code{reference}
represents (intended to be "samples" or "variants")}
}
\value{
A vector of sorted indices representing a
(possible) subset of \code{reference}
}
\description{
This utility function generates a vector of indices
corresponding to \code{reference}. When \code{ids}
or \code{idx} are provided, these are used
to subset the result.
}
\details{
\code{ids} and \code{idx} may not both be provided
}
