% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracts.R
\name{validate_tracts}
\alias{validate_tracts}
\title{Check ancestry tracts are valid}
\usage{
validate_tracts(dt_tracts)
}
\arguments{
\item{dt_tracts}{A data.table with ancestry tracts}
}
\description{
Check ancestry tracts are valid
}
