% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbm.R
\name{verify_plink}
\alias{verify_plink}
\title{Check that plink2 files exist}
\usage{
verify_plink(plink_prefix)
}
\arguments{
\item{plink_prefix}{A string with the prefix
for the plink2 file paths}
}
\value{
A list of plink2 file paths
}
\description{
Check that plink2 files exist
}
