% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAUDI.R
\name{construct_GAUDI}
\alias{construct_GAUDI}
\title{construct_GAUDI}
\usage{
construct_GAUDI(anc_FBM_obj, snps = NULL)
}
\arguments{
\item{anc_FBM_obj}{}

\item{snps}{}
}
\value{
matrix for fitting GAUDI
}
\description{
Constructs X matrix for GAUDI
}
\author{
Frank Ockerman
}
