% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaudi.R
\name{construct_gaudi}
\alias{construct_gaudi}
\title{Constructs X matrix for GAUDI}
\usage{
construct_gaudi(fbm, fbm_info, variants = NULL)
}
\arguments{
\item{fbm}{An \code{FBM.code256} object with ancestry-specific columns,
as returned by \code{make_fbm}}

\item{fbm_info}{A data.table with information on the columns in \code{fbm},
as returned by \code{make_fbm}}

\item{variants}{A character vector with IDs for variants to
include in the model}
}
\value{
The GAUDI model matrix
}
\description{
Constructs X matrix for GAUDI
}
