% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaudi.R
\name{gaudi}
\alias{gaudi}
\title{Fit GAUDI model}
\usage{
gaudi(
  fbm,
  fbm_info,
  y_train,
  ind_train = NULL,
  gamma_vec,
  k = 10,
  variants = NULL,
  verbose = FALSE,
  minlam = 0,
  maxsteps_init = 2000,
  maxsteps_cv = 2000 * 10,
  approx = FALSE
)
}
\arguments{
\item{fbm}{An \code{FBM.code256} object with ancestry-specific columns,
as returned by \code{make_fbm}}

\item{fbm_info}{A data.table with information on the columns in \code{fbm},
as returned by \code{make_fbm}}

\item{y_train}{A vector with phenotype values. If the \code{idx.train},
argument is passed, this vector must have the same length.}

\item{ind_train}{An optional vector of indices corresponding to rows
in the fbm matrix, used to subset samples for training. If provided,
y_train must be the same length and sorted to these indices.}

\item{gamma_vec}{A numeric vector with gamma values}

\item{k}{An integer with the number of cross-validation (CMSA)
folds to use for selecting gamma and lambda}

\item{variants}{A character vector with IDs for variants to
include in the model}

\item{verbose}{A boolean indicating verbosity}

\item{minlam}{A numeric with the minimum lambda value for genlasso}

\item{maxsteps_init}{An integer with the max steps for genlasso in}

\item{maxsteps_cv}{An integer with the max steps for genlasso in}

\item{approx}{A boolean indicating whether}
}
\description{
Calculates GAUDI admixture-aware polygenic scores
using a bigstatsr FBM.code256 object generated
by \code{make_fbm}.
}
\details{
This function is a re-implementation of GAUDI
provided so that it may be easily incorporated
into HAUDI workflows.
}
