% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAUDI.R
\name{GAUDI}
\alias{GAUDI}
\title{GAUDI}
\usage{
GAUDI(
  anc_FBM_obj,
  y,
  gamma_vec,
  K = 10,
  ind_train = NULL,
  snps = NULL,
  verbose = F,
  minlam = 0,
  maxsteps_init = 2000,
  maxsteps_cv = 2000 * 10
)
}
\arguments{
\item{anc_FBM_obj}{}

\item{y}{vector of responses}

\item{gamma_vec}{vector of gamma values}

\item{K}{number of folds to use in cross-validation}

\item{ind_train}{integer vector of samples to use in training model}

\item{snps}{vector of snp names to use in model}

\item{verbose}{boolean indicating verbosity}

\item{minlam}{minimum lambda value for genlasso}

\item{maxsteps_init}{max steps for genlasso in initial model}

\item{maxsteps_cv}{max steps for genlasso in cross-validation models}
}
\value{
list containing information for best cross-validated fit
}
\description{
Fit GAUDI model
}
\author{
Bryce Rowland, Frank Ockerman
}
