% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haudi.R
\name{get_beta_haudi}
\alias{get_beta_haudi}
\title{get_beta_haudi}
\usage{
get_beta_haudi(fbm_info, haudi_model)
}
\arguments{
\item{fbm_info}{data frame containing information
for FBM (chrom/pos/samples/etc.)}

\item{haudi_model}{an object of class big_sp_list,
returned by \code{haudi} or \code{lasso}}
}
\value{
An object of class \code{big_sp_list} from the \code{bigstatsr} package
}
\description{
Extract population-specific SNP coefficients from a HAUDI model
}
\author{
Frank Ockerman
}
