% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haudi.R
\name{haudi}
\alias{haudi}
\title{Fit HAUDI model}
\usage{
haudi(fbm, fbm_info, y_train, gamma_vec, family, k = 10, variants = NULL, ...)
}
\arguments{
\item{fbm}{An \code{FBM.code256} object with ancestry-specific columns,
as returned by \code{make_fbm}}

\item{fbm_info}{A data.table with information on the columns in \code{fbm},
as returned by \code{make_fbm}}

\item{y_train}{A vector with phenotype values. If the \code{idx.train},
argument is passed, this vector must have the same length.}

\item{gamma_vec}{A numeric vector of values for the gamma
tuning parameter used to control fusion in HADUI}

\item{family}{Either "gaussian" or "binomial"}

\item{k}{An integer with the number of cross-validation (CMSA)
folds to use for selecting gamma and lambda}

\item{variants}{A character vector with IDs for variants to
include in the model}
}
\value{
A list containing:
\itemize{
\item model: the final model returned by \code{bigstatsr::big_spLinReg}
or \code{bigstatsr::big_spLogReg}
\item cv_results: a data frame with mean validation loss for each
value in \code{gamma_vec}
}
}
\description{
Calculates HAUDI admixture-aware polygenic scores using
a bigstatsr FBM.code256 object generated by \code{make_fbm}.
}
\details{
This function uses bigstatsr's implementation for
sparse penalized regression and constructs a model with
appropriate multiplicative penalties for
HAUDI's admixture-aware objective function.
A sequence of gamma parameters, which control
shrinkage between ancestry-specific effects,
may be specified with \code{gamma_vec} and are
tuned using bigstatsr's Cross-Model Selection
and Averaging (CMSA). Arguments for covariates,
base predictions, etc. may be provided and are
passed to bigstatsr using \code{...}. It is assumed that the
phenotype vector y_train is sorted according
to ind.train (if this argument is passed)
or sorted by the rows of \code{fbm} (if ind.train
is not provided). This function supports
a standard Lasso model with family="gaussian"
or a penalized logistic regression model with
family="binomial".
}
