% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HAUDI.R
\name{HAUDI}
\alias{HAUDI}
\title{Train HAUDI model}
\usage{
HAUDI(anc_FBM_obj, y, gamma, ind_train, family)
}
\arguments{
\item{anc_FBM_obj}{a list containing ancestry-genotype information}

\item{y}{a vector of phenotypes}

\item{gamma}{an integer specifying the multiplicative penalty applied to ancestry-specific differences in effect}

\item{ind_train}{a vector of indices for training data}

\item{family}{either "gaussian" or "binomial"}
}
\value{
returns the model, an object of class \code{bigstatsr::big_sp_list}
}
\description{
Train HAUDI model
}
\author{
Frank Ockerman
}
