% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haudi.R
\name{lasso}
\alias{lasso}
\title{Fit ancestry-agnostic Lasso PGS model}
\usage{
lasso(fbm, fbm_info, y_train, family, k = 10, variants = NULL, ...)
}
\arguments{
\item{fbm}{An \code{FBM.code256} object with ancestry-specific columns,
as returned by \code{make_fbm}}

\item{fbm_info}{A data.table with information on the columns in \code{fbm},
as returned by \code{make_fbm}}

\item{y_train}{A vector with phenotype values. If the \code{idx.train},
argument is passed, this vector must have the same length.}

\item{family}{Either "gaussian" or "binomial"}

\item{k}{An integer with the number of cross-validation (CMSA)
folds to use for selecting gamma and lambda}

\item{variants}{A character vector with IDs for variants to
include in the model}
}
\value{
the final model returned by \code{bigstatsr::big_spLinReg}
or \code{bigstatsr::big_spLogReg}
}
\description{
This function is intended as a "null" comparator for
haudi. It calculates a standard Lasso PGS by only
including total genotypes without considering local ancestry.
}
