% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_FBM.R
\name{make_FBM}
\alias{make_FBM}
\title{Make FBM containing ancestry-genotypes and return list containing FBM and SNP info}
\usage{
make_FBM(
  vcf_file,
  FBM_pref,
  chunk_size,
  rds = NULL,
  minAC = 1,
  geno_format = "HDS",
  anc_names
)
}
\arguments{
\item{vcf_file}{File path to a VCF file}

\item{FBM_pref}{File path to store FBM. Omit the file extension}

\item{chunk_size}{Integer indicating the max number of VCF records to read at a time}

\item{rds}{Optional file path for an existing RDS  file to append data to}

\item{minAC}{Integer indicating the minimum allele count (per-ancestry) to retain}

\item{geno_format}{either "HDS" or "GT", specifying the VCF format field to use}
}
\description{
Make FBM containing ancestry-genotypes and return list containing FBM and SNP info
}
\author{
Frank Ockerman
}
