% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_FBM.R
\name{make_FBM}
\alias{make_FBM}
\title{Make FBM containing ancestry-genotypes and save in an RDS file with supplementary info}
\usage{
make_FBM(anc_vcf_file, FBM_pref, chunk_size)
}
\arguments{
\item{anc_vcf_file}{a file path to a VCF file with phased genotypes and estimated local ancestry, i.e. the output of flare (https://doi.org/10.1101/2022.08.02.502540)}

\item{FBM_pref}{a file path to store FBM and RDS file to. Omit the file extension}

\item{chunk_size}{an integer indicating the max number of VCF records to read at a time}
}
\description{
Make FBM containing ancestry-genotypes and save in an RDS file with supplementary info
}
\author{
Frank Ockerman
}
