% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbm.R
\name{make_fbm}
\alias{make_fbm}
\title{Make File-Backed Matrix input for HAUDI}
\usage{
make_fbm(
  lanc_files,
  plink_prefixes,
  fbm_prefix,
  variants = NULL,
  idx_variants_list = NULL,
  min_ac = 0,
  samples = NULL,
  idx_samples = NULL,
  anc_names = NULL,
  chunk_size = 400
)
}
\arguments{
\item{lanc_files}{A string vector with file paths for
the local ancestry input}

\item{plink_prefixes}{A string vector with the prefixes for plink2 file paths}

\item{fbm_prefix}{A string with the prefix for the
file path where a new backing file for the FBM
will be written. This prefix will be appended with ".bk"}

\item{variants}{A character vector with variant IDs to include in the FBM}

\item{idx_variants_list}{A list of integer vectors, one per plink2 input
file, containing indices of variants to include in the FBM.}

\item{min_ac}{An integer with a minimum cut off for
allele counts in each column of the FBM}

\item{samples}{A character vector of samples to
include in the FBM. It is assumed all plink2 files
have the same set of samples.}

\item{idx_samples}{An integer vector of sample indices,
corresponding to plink2 psam input, to include in the FBM.
It is assumed all plink2 files have the same set of samples.}

\item{anc_names}{A character vector containing (ordered)
names for each population in \code{ancestry_file}. If not
provided, integer values are used}

\item{chunk_size}{An integer with the max number of
variants to load at a given time while reading the
plink2 pgen file}

\item{fbm}{An existing \code{FBM.code256} object to which
results are appended.}
}
\value{
A list containing:
\code{fbm}: an \code{FBM.code256} object from \code{bigstatsr},
which links to a file-backed matrix used for
fittting HAUDI models. Each column contains
either ancestry-specific genotypes (e.g. number of
alternate alleles where ancestry is also "pop0") or
the total genotype. This object also contains a
\code{samples} attribute ordered according to the samples
in the matrix rows.
\code{info} a \code{data.table} containing information
about columns in the FBM.
}
\description{
Processes .lanc local ancestry files and corresponding plink2 files,
producing a file-backed matrix object and associated
information used for HAUDI.
}
\details{
Only one (or neither) of \code{variants} and \code{idx_variants_list} may be provided.
\code{variants} is a single character with all variant IDs (across input files),
while \code{idx_variants_list} is a list where each element is a vector of indices
corresponding to a single set of plink2 files (e.g. each vector is
per-chromosome). Only one (or neither) of \code{samples} and \code{idx_samples} may
be provided. Exactly one of \code{fbm_prefix} or \code{fbm} may be provided.
It is assumed that the local ancestry input files and plink2 inputs
are sorted by chromosome and position.
}
