% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fbm.R
\name{make_fbm}
\alias{make_fbm}
\title{Make HAUDI-style FBM and return list with FBM and SNP info}
\usage{
make_fbm(
  vcf_file,
  fbm_pref,
  chunk_size,
  rds = NULL,
  min_ac = 1,
  geno_format = "HDS",
  anc_names
)
}
\arguments{
\item{vcf_file}{File path to a VCF file}

\item{fbm_pref}{File path to store FBM. Omit the file extension}

\item{chunk_size}{The maximum number of VCF records to read at a time}

\item{rds}{Optional file path for an existing RDS  file to append data to}

\item{min_ac}{Minimum allele count (per-ancestry) to retain}

\item{geno_format}{The VCF format field to use ("HDS" or "GT")}

\item{anc_names}{vector of labels for each population}
}
\description{
Make HAUDI-style FBM and return list with FBM and SNP info
}
