% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbm.R
\name{make_haudi_chunk}
\alias{make_haudi_chunk}
\title{Produces HAUDI matrix and variant info for a single chunk}
\usage{
make_haudi_chunk(chunk, pgen, pvar, tracts, min_ac, anc_names, idx_samples)
}
\arguments{
\item{chunk}{An integer vector of indices in the pgen file}

\item{pgen}{A \code{pgen} object produced by \code{pgenlibr::NewPgen()}}

\item{pvar}{A data frame corresponding to the pvar file}

\item{tracts}{A list of data.frames, where each element corresponds to a
sample and has columns \code{index} (representing indices
in a plink2 pvar file) and "anc0" and "anc1", corresponding to ancestries for
each haplotype}

\item{min_ac}{An integer with the minimum allele count to
retain columns in the HAUDI matrix}

\item{anc_names}{An ordered vector of ancestry names}

\item{idx_samples}{An ordered integer vector of samples to
retain from the pgen}
}
\value{
A list with the HAUDI matrix and variant info for the chunk
}
\description{
Produces HAUDI matrix and variant info for a single chunk
}
