% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracts.R
\name{read_lanc}
\alias{read_lanc}
\title{Read .lanc file into list of tracts}
\usage{
read_lanc(lanc_file)
}
\arguments{
\item{lanc_file}{A string with the file path to a .lanc file from \code{Admix-kit}}
}
\value{
A flattened list structure with elements:
left_haps: ancestry for hap0 (integer) across samples
right_haps: ancestry for hap1 (integer) across samples
breakpoints: ancestry tract breakpoints (idx in pvar file) across samples
offsets: cumulative end indices separating samples
}
\description{
Read .lanc file into list of tracts
}
