% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracts.R
\name{read_lanc}
\alias{read_lanc}
\title{Read .lanc file into list of tracts}
\usage{
read_lanc(lanc_file)
}
\arguments{
\item{lanc_file}{A string with the file path to a .lanc file from \code{Admix-kit}}
}
\value{
A list of data.frames, where each element corresponds to a sample,
and has columns \code{index} (representing indices
in a plink2 pvar file) and "anc0" and "anc1", corresponding to ancestries for
each haplotype.
}
\description{
Read .lanc file into list of tracts
}
